fix lp3.jar import not working for lp2
