+++ package/android/src/lp2/java/com/mrousavy/camera/core/CameraSession+EIS.kt
@@ -1,0 +1,5 @@
+package com.mrousavy.camera.core
+
+fun CameraSession.setEISMode(newMode: String) {
+  return
+}
